# backend/docs/swagger.yaml
basePath: /
definitions:
...
  billing.Payment:
    properties:
      amount:
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        type: string
      stripePaymentId:
        type: string
    type: object
  billing.Plan:
    properties:
      amount:
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      description:
        type: string
      features:
        items:
          type: string
        type: array
      id:
        type: string
      interval:
        type: string
      isActive:
        type: boolean
      isPopular:
        type: boolean
      name:
        type: string
      stripePriceId:
        type: string
      stripeProductId:
        type: string
      updatedAt:
        type: string
    type: object
  billing.Subscription:
    properties:
      cancelAtPeriodEnd:
        type: boolean
      createdAt:
        type: string
      currentPeriodEnd:
        type: string
      currentPeriodStart:
        type: string
      id:
        type: string
      plan:
        $ref: '#/definitions/billing.Plan'
      planId:
        type: string
      status:
        type: string
      stripeCustomerId:
        type: string
      stripeSubscriptionId:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  httpx.Error:
    properties:
      detail:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: LHCHub home-services marketplace API (Go backend). Clients post jobs,
    vetted providers apply, and payment is held in escrow.
  termsOfService: http://swagger.io/terms/
  title: LHCHub API
  version: "1.0"
paths:
...
  /api/billing/cancel/:
    post:
      consumes:
      - application/json
      description: Schedules the current subscription to cancel at period end.
      parameters:
      - description: '{subscriptionId}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "404":
          description: Not Found
          schema:
            type: object
        "502":
          description: Bad Gateway
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Cancel subscription
      tags:
      - billing
  /api/billing/checkout/:
    post:
      consumes:
      - application/json
      description: Creates a Stripe Checkout session for a plan and returns its URL.
      parameters:
      - description: '{planId}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "502":
          description: Bad Gateway
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create checkout session
      tags:
      - billing
  /api/billing/payments/:
    get:
      description: Returns the current user's payments.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/billing.Payment'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: List payments
      tags:
      - billing
  /api/billing/plans/:
    get:
      description: Returns active billing plans ordered by amount.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/billing.Plan'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: List plans
      tags:
      - billing
  /api/billing/portal/:
    post:
      consumes:
      - application/json
      description: Creates a Stripe Billing Portal session and returns its URL.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "502":
          description: Bad Gateway
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create billing portal session
      tags:
      - billing
  /api/billing/resume/:
    post:
      consumes:
      - application/json
      description: Resumes a subscription scheduled to cancel.
      parameters:
      - description: '{subscriptionId}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "404":
          description: Not Found
          schema:
            type: object
        "502":
          description: Bad Gateway
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Resume subscription
      tags:
      - billing
  /api/billing/subscriptions/:
    get:
      description: Returns the current user's subscriptions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/billing.Subscription'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: List subscriptions
      tags:
      - billing
  /api/billing/subscriptions/current/:
    get:
      description: Returns the current user's active subscription (or null).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: Current subscription
      tags:
      - billing
  /api/billing/verify-session/:
    get:
      description: Verifies a Stripe Checkout session and syncs subscription/payment
        rows.
      parameters:
      - description: Stripe Checkout session id
        in: query
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "502":
          description: Bad Gateway
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Verify checkout session
      tags:
      - billing
  /api/billing/webhook/:
    post:
      consumes:
      - application/json
      description: Receives Stripe events and syncs subscription/payment rows.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: Stripe webhook
      tags:
      - billing
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"